refactor(modal): improve scalability and maintainability

+++ frontend/src/components/common/modal/content/MemberAction.tsx
@@ -3,27 +3,32 @@ import Button from "../../Button";
 import TextArea from "../../TextArea";
 import { type User } from "../../../../types/user";
 import { useParams } from "react-router-dom";
-import { useMembers } from "../../../../hooks/community/useMembers";
+import { useMembers } from "../../../../contexts/community/useMembers";
 
-type KickUserProps = {
+type MemberActionProps = {
     onClose: () => void;
     user?: User;
+    action: "kick" | "ban";
 }
 
-export default function KickUser({ user, onClose }: KickUserProps) {
+const config = {
+    kick: { title: "Kick", endpoint: "kick", description: "Are you sure you want to kick this user?" },
+    ban: { title: "Ban", endpoint: "ban", description: "Are you sure you want to ban this user?" },
+};
+
+export default function MemberAction({ user, onClose, action }: MemberActionProps) {
     const token = localStorage.getItem("token");
     const { communityId } = useParams();
     const apiUrl = window.GLOBAL_ENV.API_ENDPOINT;
     const { setMembers } = useMembers();
     const [reason, setReason] = useState("");
+    const { title, endpoint, description } = config[action];
 
-    const handleKick = async () => {
-        if (!user) return;
-        if (!token) return;
-        if (!communityId) return;
+    const handleAction = async () => {
+        if (!user || !token || !communityId) return;
         try {
             const response = await fetch(
-                `${apiUrl}/api/community/${communityId}/members/${user.id}/kick`,
+                `${apiUrl}/api/community/${communityId}/members/${user.id}/${endpoint}`,
                 {
                     method: "DELETE",
                     headers: {
@@ -38,15 +43,15 @@ export default function KickUser({ user, onClose }: KickUserProps) {
                 onClose();
             }
         } catch (error) {
-            console.log(`error kicking ${user.username}: `, error);
+            console.log(`error ${action}ing ${user.username}: `, error);
         }
     };
 
     return (
         <>
             <div className="flex flex-col text-center">
-                <div className="font-bold text-[20px]">Kick {user?.username}</div>
-                <div className="text-gray-500">Are you sure you want to kick this user?</div>
+                <div className="font-bold text-[20px]">{title} {user?.username}</div>
+                <div className="text-gray-500">{description}</div>
             </div>
             <TextArea
                 placeholder="Reason?"
@@ -54,7 +59,7 @@ export default function KickUser({ user, onClose }: KickUserProps) {
                 onChange={(e) => setReason(e.target.value)}
                 maxLength={100}
             />
-            <Button text="Kick" isDanger onClick={handleKick} />
+            <Button text={title} isDanger onClick={handleAction} />
         </>
     );
 }
